pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ado-readiness"
version = "0.1.0"
description = "Azure DevOps Migration Readiness Analyzer - Pre-migration assessment tool for ADO to GitHub migrations"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "nkusakula" }
]
keywords = ["azure-devops", "github", "migration", "assessment", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Build Tools",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "httpx>=0.25.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
ado-readiness = "ado_readiness.cli:app"

[project.urls]
Homepage = "https://github.com/nkusakula/ADO-to-GitHub-Pre-migration-assessment-tool"
Repository = "https://github.com/nkusakula/ADO-to-GitHub-Pre-migration-assessment-tool"
Issues = "https://github.com/nkusakula/ADO-to-GitHub-Pre-migration-assessment-tool/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/ado_readiness"]

[tool.ruff]
target-version = "py310"
line-length = 100
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.10"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
